@@ -7,7 +7,7 @@ publish_to: 'none'
 
 environment:
   sdk: '>=3.6.0 <4.0.0'
-  flutter: ">=3.27.1"
+  flutter: ">=3.29.0"
 
 
 dependencies:
@@ -16,7 +16,7 @@ dependencies:
 
   cupertino_icons: ^1.0.8
 
-  flutter_badge_manager:
+  flutter_badge_manager_android:
     path: ../
 
 
@@ -29,7 +29,7 @@ dev_dependencies:
     sdk: flutter
 
   # Linting
-  flutter_lints: ^5.0.0
+  flutter_lints: ^6.0.0
 
 
 flutter:
